requirements.txt
redis>=5.0.0
numpy>=1.23
sentence-transformers>=2.2.2
torch>=2.0.0
streamlit>=1.28.0
psutil>=5.9.5
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=9.5.0
regex>=2023.10.3